(PC-24426)[PRO] refactor: cleanup booking table cells
