+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/Cells/BeneficiaryCell.module.scss
