+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/Cells/BookingIsDuoCell.module.scss
