+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/Cells/BookingOfferCell.module.scss
