+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/Cells/BookingStatusCell.module.scss
