+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/Cells/CellFormatter.module.scss
