+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/Cells/NumberOfTicketsAndPriceCell.tsx
@@ -1,0 +1,23 @@
+import React from 'react'
+
+import { CollectiveBookingResponseModel } from 'apiClient/v1'
+import { pluralizeString } from 'utils/pluralize'
+
+interface NumberOfTicketsAndPriceCellProps {
+  booking: CollectiveBookingResponseModel
+}
+
+export const NumberOfTicketsAndPriceCell = ({
+  booking,
+}: NumberOfTicketsAndPriceCellProps) => {
+  const numberOfTickets = booking.stock.numberOfTickets
+
+  return (
+    <div>
+      <div>
+        {numberOfTickets} {pluralizeString('place', numberOfTickets)}
+      </div>
+      <div>{booking.bookingAmount}€</div>
+    </div>
+  )
+}
