+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/Cells/__specs__/BookingTokenCell.spec.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import React from 'react'
 
-import BookingTokenCell, { BookingTokenCellProps } from '../BookingTokenCell'
+import { BookingTokenCell, BookingTokenCellProps } from '../BookingTokenCell'
 
 const renderBookingTokenCell = (props: BookingTokenCellProps) =>
   render(<BookingTokenCell {...props} />)
