+++ pro/src/screens/Bookings/BookingsRecapTable/BookingsTable/Cells/index.ts
@@ -1,0 +1,10 @@
+export { BeneficiaryCell } from './BeneficiaryCell'
+export { BookingDateCell } from './BookingDateCell'
+export { BookingIdCell } from './BookingIdCell'
+export { BookingIsDuoCell } from './BookingIsDuoCell'
+export { BookingOfferCell } from './BookingOfferCell'
+export { BookingStatusCell } from './BookingStatusCell'
+export { BookingTokenCell } from './BookingTokenCell'
+export { NumberOfTicketsAndPriceCell } from './NumberOfTicketsAndPriceCell'
+export { DetailsButtonCell } from './DetailsButtonCell'
+export { InstitutionCell } from './InstitutionCell'
