--- pro/src/screens/Bookings/BookingsRecapTable/components/CellsFormatter/index.ts
@@ -1,9 +1,0 @@
-export { default as BeneficiaryCell } from './BeneficiaryCell'
-export { default as BookingDateCell } from './BookingDateCell'
-export { default as BookingIsDuoCell } from './BookingIsDuoCell'
-export { default as BookingOfferCell } from './BookingOfferCell'
-export { default as BookingStatusCell } from './BookingStatusCell'
-export { default as BookingTokenCell } from './BookingTokenCell'
-export { default as NumberOfTicketsAndPriceCell } from './NumberOfTicketsAndPriceCell'
-export { default as DetailsButtonCell } from './DetailsButtonCell'
-export { default as InstitutionCell } from './InstitutionCell'
